LayoutTests/fast/scrolling/arrow-key-scroll-in-rtl-document-expected.txt

PASS: scrollLeft is less than or equal to -40
